Refactor fusion tab logic into dedicated module
